兼容性问题调整。

+++ spring-boot-samples-parent/spring-boot-samples-sdk-parent/spring-boot-sample-sdk-dwz/src/test/java/icu/easyj/spring/boot/sample/sdk/dwz/baidu/restcontroller/BaiduDwzTemplateControllerTest.java
@@ -42,7 +42,7 @@ class BaiduDwzTemplateControllerTest extends BaseSpringBootMockMvcTest {
 				.characterEncoding().is(StandardCharsets.UTF_8)
 				.content(DwzResponse.class).is(resp -> {
 					Assertions.assertTrue(resp.getShortUrl().startsWith("https://dwz.cn/"));
-					Assertions.assertEquals(0L, resp.getExpireIn());
+					Assertions.assertEquals(0L, resp.getExpireIn().longValue());
 				});
 	}
 }
